# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Identity Auth API",
    "description": "Kata de autenticación con OCR, liveness y face match usando FastAPI.",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/session": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Start Authentication",
        "description": "Crea una nueva sesión de autenticación y la devuelve al cliente.",
        "operationId": "start_authentication_auth_session_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/document": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Upload Document",
        "description": "Recibe la imagen del documento asociada a una sesión\ny delega el flujo completo al document_service:\n  - validación de sesión\n  - validación de imagen\n  - OCR\n  - validación contra repositorio local\n  - actualización de sesión",
        "operationId": "upload_document_auth_document_post",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sessionid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_auth_document_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/selfie": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Upload Selfie",
        "description": "Recibe la selfie asociada a una sesión de autenticación\ny delega el flujo al selfie_service:\n  - validación de sesión\n  - validación de imagen\n  - análisis de liveness\n  - actualización de sesión",
        "operationId": "upload_selfie_auth_selfie_post",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sessionid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_selfie_auth_selfie_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/decision": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Finalize Decision",
        "description": "Toma la decisión final de autenticación usando el decision_service:\n  - revisa flags de flujo (document/selfie procesados)\n  - valida documento, OCR y liveness\n  - actualiza el estado de la sesión (APPROVED / REJECTED)\n  - devuelve un DecisionResult con todos los detalles",
        "operationId": "finalize_decision_auth_decision_post",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sessionid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/session-debug": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Session Debug",
        "description": "Endpoint de apoyo para ver el estado crudo de una sesión.\n(solo para debugging / demo).",
        "operationId": "session_debug_auth_session_debug_get",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sessionid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_auth_document_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_auth_document_post"
      },
      "Body_upload_selfie_auth_selfie_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_selfie_auth_selfie_post"
      },
      "DecisionResult": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "documentValid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentvalid"
          },
          "documentValidationReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentvalidationreason"
          },
          "ocrConfidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocrconfidence"
          },
          "captureQuality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capturequality"
          },
          "livenessScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Livenessscore"
          },
          "faceMatchScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facematchscore"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "status"
        ],
        "title": "DecisionResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "StartAuthResponse": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "message",
          "createdAt"
        ],
        "title": "StartAuthResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}